refactor: tidy project structure
- tsconfig を 1 本に統合（app/node 分離をやめて単一の tsconfig.json で完結）
- src/lib/KeiLisp/Graphist.js を src/ 配下から attic/ に退避
  （ビルド対象外であることを明示。将来の別パッケージ移植用に保持）
- src/repl-bus.ts を Vue 慣習に沿って分割
  - src/composables/useReplOutput.ts: 共有 ref を提供する composable
  - src/shims/process.ts: process.stdout/stderr の差し替えロジック
- env.d.ts から冗長な triple-slash reference を削除
- build script を vue-tsc -b から --noEmit に変更（project references 廃止に追従）
