+++ env.d.ts
@@ -1,5 +1,3 @@
-/// <reference types="vite/client" />
-
 declare module '*.vue' {
   import type { DefineComponent } from 'vue';
   const component: DefineComponent<Record<string, never>, Record<string, never>, unknown>;
