+++ src/composables/useReplOutput.ts
@@ -1,0 +1,5 @@
+import { ref, type Ref } from 'vue';
+
+const replOutput = ref('');
+
+export const useReplOutput = (): Ref<string> => replOutput;
